Create ProximityManager.cs
The second version of the Unity Crowds tutorial for part 2. I am creating a new project instead of recommitting the existing one for the purposes of new people who want to do the first part of the video and follow along. If I gave them the updated one it would break things. Thanks so much for having interest in my work. I really appreciate it.